% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allometry.R
\name{dbh2bl}
\alias{dbh2bl}
\title{Dbh to leaf biomass allometry}
\usage{
dbh2bl(dbh,ipft)
}
\arguments{
\item{dbh}{diameter at breast height}

\item{ipft}{PFT}
}
\value{
bleaf leaf biomass
}
\description{
This is an allometric function to return the leaf
biomass as a function of dbh
}

