% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allometry.R
\name{dbh2h}
\alias{dbh2h}
\title{dbh to height allometry}
\usage{
dbh2h(dbh,ipft)
}
\arguments{
\item{dbh}{diameter at breast height}

\item{ipft}{PFT}
}
\value{
h height
}
\description{
This is an allometric function to return the height as a function of dbh
}

