% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allometry.R
\name{dbh2wai}
\alias{dbh2wai}
\title{dbh to wood area index allometry}
\usage{
dbh2wai(dbh,ipft)
}
\arguments{
\item{dbh}{diameter at breast height}

\item{ipft}{PFT}

\item{chambers}{Use Chambers method?(set to FALSE)}
}
\value{
wai wood area index
}
\description{
This is an allometric function to return the wood area index
it is derived from Ahrends et al. (2010).
}

