% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.all.manfredo.R
\docType{data}
\name{file.dir}
\alias{file.dir}
\title{Main plot utility}
\format{An object of class \code{character} of length 1.}
\usage{
file.dir
}
\author{
Marcos Longo & Manfredo di Porcia
}
\keyword{datasets}

