% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allometry.R
\name{h2dbh}
\alias{h2dbh}
\title{Height to dbh allometry}
\usage{
h2dbh(h,ipft)
}
\arguments{
\item{h}{height}

\item{ipft}{PFT}
}
\value{
dbh diameter at breast height
}
\description{
This is an allometric function to return the dbh as a function of height
}

