% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourcer.R
\name{sourcer}
\alias{sourcer}
\title{Sourcer is a function to source the necessary files in the main plottin script}
\usage{
sourcer(source.dir)
}
\arguments{
\item{source.dir}{directory to source}
}
\description{
Sourcer is a function to source the necessary files in the main plottin script
}

